python/pyproject.toml
[tool.poetry]
name = "python-monorepo"
version = "0.1.0"
description = "manage dependencies for Python projects"
license = "MIT"
authors = ["genkey6 <genkey6@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
starlette = "^0.23.0"
gunicorn = "^20.1.0"
uvicorn = "^0.20.0"
python-multipart = "^0.0.5"
strawberry-graphql = {extras = ["asgi"], version = "^0.151.0"}

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
black = "^22.6.0"
flake8 = "^6.0.0"
invoke = "^1.7.1"
isort = "^5.10.1"
mypy = "^0.991"
pytest = "^7.1.2"
pyproject-flake8 = "^6.0.0.post1"
types-invoke = "^1.7.3.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules -n auto"
doctest_optionflags = [
  "NORMALIZE_WHITESPACE",
  "IGNORE_EXCEPTION_DETAIL",
]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
profile = "black"
use_parentheses = true

[tool.flake8]
max-line-length = 120
max-complexity = 18
select = "B,C,E,F,N,W,T4"
# E203/W503 conflict with black. For details see here:
# https://black.readthedocs.io/en/stable/faq.html#why-are-flake8-s-e203-and-w503-violated
extend-ignore = "E203, W503"
# Ignore 'imported but unused' in __init__.py
per-file-ignores = "__init__.py:F401"
exclude = ".git,__pycache__,old,build,dist,.venv,.mypy_cache,.pytest_cache"
extend-exclude = "venv"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
check_untyped_defs = true
show_column_numbers = true
warn_unused_configs = true
warn_unused_ignores = true
namespace_packages = true
exclude = "tasks"
plugins = "strawberry.ext.mypy_plugin"
